requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
pydantic
python-multipart

langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
anthropic
passlib[bcrypt]
python-jose[cryptography]
requests
elevenlabs==0.2.26
